+++ src/main/java/com/enginemachining/capabilities/storage/EnergyStorage.java
@@ -14,7 +14,7 @@ public class EnergyStorage implements Capability.IStorage<IEnergyHandler> {
     @Override
     public INBT writeNBT(Capability<IEnergyHandler> capability, IEnergyHandler instance, Direction side) {
         CompoundNBT nbt = new CompoundNBT();
-        nbt.putInt("energy", instance.getStoredPower());
+        nbt.putFloat("energy", instance.getStoredPower());
         return nbt;
     }
 
@@ -22,6 +22,6 @@ public class EnergyStorage implements Capability.IStorage<IEnergyHandler> {
     public void readNBT(Capability<IEnergyHandler> capability, IEnergyHandler instance, Direction side, INBT nbt) {
         CompoundNBT e = (CompoundNBT) nbt;
         if(!(instance instanceof DefaultEnergyHandler)) throw new IllegalArgumentException("Cannot deserialize a non default implementation.");
-        ((DefaultEnergyHandler)instance).energy = e.getInt("energy");
+        ((DefaultEnergyHandler)instance).energy = e.getFloat("energy");
     }
 }
